# docs/swagger.yaml
basePath: /api/v1/cbesuperapp/cps_action
definitions:
...
  cbe-super-app-cps-action_internal_constants_dto_mini_app_merchant.KYCDTO:
    properties:
      representative:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_mini_app_merchant.RepresentativeDTO'
      status:
        type: string
    type: object
  cbe-super-app-cps-action_internal_constants_dto_mini_app_merchant.MiniAppMerchantDTO:
    properties:
      account_number:
        type: string
      email:
        type: string
      merchant_name:
        type: string
      merchant_representative_name:
        type: string
      phone_number:
        type: string
      type:
        type: string
    type: object
  cbe-super-app-cps-action_internal_constants_dto_mini_app_merchant.MiniAppMerchantResponseDTO:
    properties:
      account_number:
        type: string
      code:
        type: string
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      is_deleted:
        type: boolean
      kyc:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_mini_app_merchant.KYCDTO'
      last_modified:
        type: string
      merchant_name:
        type: string
      mini_apps:
        items:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_model.MiniApps'
        type: array
      type:
        type: string
    type: object
  cbe-super-app-cps-action_internal_constants_dto_mini_app_merchant.PaginatedMiniAppResponseResponse:
    properties:
      docs:
        items:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_mini_app_merchant.MiniAppMerchantResponseDTO'
        type: array
      meta:
        $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_types.PaginationMeta'
    type: object
  cbe-super-app-cps-action_internal_constants_dto_mini_app_merchant.RepresentativeDTO:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
...
  cbe-super-app-cps-action_internal_constants_localization.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      ok:
        type: boolean
      status:
        type: integer
      timestamp:
        type: string
    type: object
...
  cbe-super-app-cps-action_internal_constants_model.MiniApps:
    properties:
      enabled:
        type: boolean
      id:
        type: string
      is_deleted:
        type: boolean
    type: object
...
  cbe-super-app-cps-action_internal_constants_types.PaginationMeta:
    properties:
      has_next_page:
        type: boolean
      has_prev_page:
        type: boolean
      limit:
        type: integer
      next_page:
        type: integer
      page:
        type: integer
      paging_counter:
        type: integer
      prev_page:
        type: integer
      total_docs:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: contact@eaglelionsystems.com
    name: API Support
  description: API documentation for CPS Action service
  title: CPS Action API
  version: 1.0.0
paths:
...
  /mini-app-merchants:
    get:
      description: Retrieves a paginated list of mini app merchants
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_mini_app_merchant.PaginatedMiniAppResponseResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: List Mini App Merchants
      tags:
      - MiniAppMerchant
    post:
      consumes:
      - application/json
      description: Creates a new mini app merchant
      parameters:
      - description: Mini App Merchant DTO
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_mini_app_merchant.MiniAppMerchantDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_mini_app_merchant.MiniAppMerchantDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Create Mini App Merchant
      tags:
      - MiniAppMerchant
  /mini-app-merchants/{id}:
    delete:
      description: Deletes a mini app merchant by ID
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Delete Mini App Merchant
      tags:
      - MiniAppMerchant
    get:
      description: Retrieves a mini app merchant by ID
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_mini_app_merchant.MiniAppMerchantResponseDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Get Mini App Merchant by ID
      tags:
      - MiniAppMerchant
    put:
      consumes:
      - application/json
      description: Updates an existing mini app merchant
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: string
      - description: Mini App Merchant DTO
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_dto_mini_app_merchant.MiniAppMerchantDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Update Mini App Merchant
      tags:
      - MiniAppMerchant
  /mini-app-merchants/{id}/disable:
    patch:
      description: Disables a mini app merchant by ID
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Disable Mini App Merchant
      tags:
      - MiniAppMerchant
  /mini-app-merchants/{id}/enable:
    patch:
      description: Enables a mini app merchant by ID
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/cbe-super-app-cps-action_internal_constants_localization.StandardResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Enable Mini App Merchant
      tags:
      - MiniAppMerchant
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"